Fix textinput animation when keyboard changes its size
When switching from "normal keyboard" to "emoji keyboard" the keyboard height changes without an animation. Duration in this case is set to 0. Because we're still using an animation here, the textinput position "lags", which results in visual glitches (especially when the new keyboard height is smaller we'll see a black background).